bug fix for data check

+++ build_train_model_v2.py
@@ -9,6 +9,7 @@ import os
 import numpy as np
 import pathlib
 import tensorflow as tf
+import shutil
 from tensorflow.keras.optimizers import Adam, RMSprop, SGD
 from tensorflow.keras.callbacks import ModelCheckpoint, LearningRateScheduler
 from tensorflow.keras.callbacks import ReduceLROnPlateau
@@ -120,11 +121,12 @@ if __name__ == '__main__':
         print("No of incompletely extracted motion stream examples:",len(incomp_motion))
         print("Droping incomplete examples")
         for folder in incomp_appearance:
-            os.remove(os.path.join(appearance_path,folder))
-            os.remove(os.path.join(motion_path,folder))
+            shutil.rmtree(os.path.join(appearance_path,folder))
+            os.rmtree(os.path.join(motion_path,folder))
         for folder in incomp_motion:
-            os.remove(os.path.join(appearance_path,folder))
-            os.remove(os.path.join(motion_path,folder))
+            if folder not in appearance_path:
+                os.remove(os.path.join(appearance_path,folder))
+                os.remove(os.path.join(motion_path,folder))
         ap_list = os.listdir(appearance_path)
         print("No of data folders :{} Appearance ,{} Motion".format(len(appearance_path),len(motion_path)))
     ## Check for txt file and tfrecord paths
